python-requirements.txt
Django==1.8.6
MySQL-python==1.2.5
django-bootstrap3==6.2.2
django-tagging==0.4
django-url-tools==0.0.8
